# api-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the release candidate version of the \"Swiss Corporate API\"\
    \ specification.\n"
  version: "1.0.0.2-SNAPSHOT"
  title: "Swiss Corporate API"
  contact:
    name: "SIX Group"
    url: "https://www.six-group.com/"
    email: "swisscorpapi@six-group.com"
host: "tbd.six-group.com"
basePath: "/api/bankingservices/corporate/v1"
...
schemes:
- "https"
paths:
  /payments:
    post:
      tags:
      - "payments"
      summary: "Submit payment instructions"
      operationId: "submitPaymentInstructions"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      - name: "X-PSU-IP-Address"
        in: "header"
        description: "IP address of the user initiating the operation"
        required: true
        type: "string"
      - name: "X-PSU-User-Agent"
        in: "header"
        description: "User agent of the user initiating the operation"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Payment instruction details as defined by data model.\n"
        required: true
        schema:
          $ref: "#/definitions/paymentSubmissionRequest"
      responses:
        201:
          description: "Created."
          headers:
            Location:
              type: "string"
              description: "Location of the submitted payment submission message.\
                \ Add /status for the status report."
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Payments"
  /payments/{submissionId}:
    get:
      tags:
      - "payments"
      summary: "Retrieve a payment submission"
      operationId: "retrievePaymentSubmission"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "submissionId"
        in: "path"
        description: "id of payment submission"
        required: true
        type: "string"
        maxLength: 30
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      responses:
        200:
          description: "The original payment submission"
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
          schema:
            $ref: "#/definitions/paymentSubmissionRequest"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Payments"
  /payments/{submissionId}/status:
    get:
      tags:
      - "payments"
      summary: "Retrieve payment submission status information"
      operationId: "retrievePaymentSubmissionStatus"
      produces:
      - "application/json"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Bearer followed by a base64 encoded OAuth access token"
        required: true
        type: "string"
      - name: "submissionId"
        in: "path"
        description: "id of payment submission"
        required: true
        type: "string"
        maxLength: 30
      - name: "X-CorAPI-Target-ID"
        in: "header"
        description: "ID that identifies the provider (e.g., a financial institution).\
          \ (SCOPE: SIX)"
        required: true
        type: "string"
      - name: "X-CorAPI-Client-ID"
        in: "header"
        description: "ID of the client forwarded to the provider. (SCOPE: FI)"
        required: true
        type: "string"
      - name: "X-Correlation-ID"
        in: "header"
        description: "Unique ID (defined by the caller) which will be reflected back\
          \ in the response."
        required: true
        type: "string"
      - name: "User-Agent"
        in: "header"
        description: "Name and version of the of the Client software"
        required: true
        type: "string"
      responses:
        200:
          description: "Status information for the payment submission"
          headers:
            X-Correlation-ID:
              type: "string"
              description: "Reflects the ID (set by the caller) from the request."
          schema:
            $ref: "#/definitions/paymentSubmissionStatus"
        400:
          description: "Bad Request - The format of the request was invalid."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        401:
          description: "Unauthorized - Either no token or an invalid (e.g., token\
            \ expired) token was received."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        403:
          description: "Forbidden - A valid OAuth Token was received, but access was\
            \ denied."
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        404:
          description: "Not Found - Either the endpoint does not exist or a requested\
            \ resource is not yet available (e.g., account statements)"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        405:
          description: "Method Not Allowed"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
        500:
          description: "Internal Server Error"
          headers:
            Content-Type:
              type: "string"
              description: "application/problem+json according to RFC7807"
            Content-Encoding:
              type: "string"
              description: "always UTF-8"
            Content-Language:
              type: "string"
              description: "always en"
            X-Correlation-ID:
              type: "string"
              description: "Reflected client defined ID from request."
          schema:
            $ref: "#/definitions/commonErrorResponse"
      x-swagger-router-controller: "Payments"
...
definitions:
  paymentSubmissionRequest:
    type: "object"
    required:
    - "bookingInstruction"
    - "debtorAccount"
    - "initiatingPartyId"
    - "messageId"
    - "requestedExecutionDate"
    - "transactions"
    properties:
      messageId:
        type: "string"
        example: "eb6305c91f7f49deaed016487c27b42d"
        maxLength: 35
      initiatingPartyId:
        type: "string"
        example: "TPP01746"
        maxLength: 32
      requestedExecutionDate:
        type: "string"
        format: "date"
        example: "2018-04-07"
      debtorAccount:
        $ref: "#/definitions/paymentIbanAccount"
      bookingInstruction:
        $ref: "#/definitions/paymentBookingInstruction"
      transactions:
        type: "array"
        items:
          $ref: "#/definitions/paymentInstructionItem"
    title: "Payment Submission Request"
    example:
      initiatingPartyId: "TPP01746"
      bookingInstruction: "BATCHBOOKING_SALA_CND"
      debtorAccount:
        identification: "CH9300762011623852957"
        type: "IBAN"
      requestedExecutionDate: "2018-04-07"
      messageId: "eb6305c91f7f49deaed016487c27b42d"
      transactions:
      - otherDetails:
          chargeBearer: "SHAR"
          creditorAgent:
            clearingSystemMemberIdentification:
              code: "CHBCC"
              memberId: "00230"
            bic: "BDEMMXMM"
          creditorAccount:
            identification: "762011623852957"
            type: "OTHER"
          creditor:
            postalAddress:
              unstructured:
                country: "CH"
                addressLines:
                - "Robert Schneider SA"
                - "Rue de la gare 24"
              structured:
                country: "CH"
                streetName: "Rue de la gare"
                townName: "Biel"
                buildingNumber: "24"
                postCode: "2501"
            name: "name"
          remittanceReference:
            reference: "210000000003139471430009017"
            type: "SCOR"
        ibanDetails:
          chargeBearer: "SHAR"
          creditorAgent:
            clearingSystemMemberIdentification:
              code: "CHBCC"
              memberId: "00230"
            bic: "BDEMMXMM"
          remittanceInformation: "Verguetung Juli"
          creditorAccount:
            identification: "CH9300762011623852957"
            type: "IBAN"
          sepaIndicator: true
          creditor:
            postalAddress:
              unstructured:
                country: "CH"
                addressLines:
                - "Robert Schneider SA"
                - "Rue de la gare 24"
              structured:
                country: "CH"
                streetName: "Rue de la gare"
                townName: "Biel"
                buildingNumber: "24"
                postCode: "2501"
            name: "name"
          remittanceReference:
            reference: "210000000003139471430009017"
            type: "SCOR"
        instructionId: "INSTR-001"
        instructedAmount:
          amount: "10.25"
          currency: "CHF"
        endToEndId: "ENDTOENDID-001"
        isrDetails:
          creditorAccount:
            identification: "762011623852957"
            type: "OTHER"
          creditor:
            postalAddress:
              unstructured:
                country: "CH"
                addressLines:
                - "Robert Schneider SA"
                - "Rue de la gare 24"
              structured:
                country: "CH"
                streetName: "Rue de la gare"
                townName: "Biel"
                buildingNumber: "24"
                postCode: "2501"
            name: "name"
          remittanceReference:
            reference: "210000000003139471430009017"
            type: "ISR"
      - otherDetails:
          chargeBearer: "SHAR"
          creditorAgent:
            clearingSystemMemberIdentification:
              code: "CHBCC"
              memberId: "00230"
            bic: "BDEMMXMM"
          creditorAccount:
            identification: "762011623852957"
            type: "OTHER"
          creditor:
            postalAddress:
              unstructured:
                country: "CH"
                addressLines:
                - "Robert Schneider SA"
                - "Rue de la gare 24"
              structured:
                country: "CH"
                streetName: "Rue de la gare"
                townName: "Biel"
                buildingNumber: "24"
                postCode: "2501"
            name: "name"
          remittanceReference:
            reference: "210000000003139471430009017"
            type: "SCOR"
        ibanDetails:
          chargeBearer: "SHAR"
          creditorAgent:
            clearingSystemMemberIdentification:
              code: "CHBCC"
              memberId: "00230"
            bic: "BDEMMXMM"
          remittanceInformation: "Verguetung Juli"
          creditorAccount:
            identification: "CH9300762011623852957"
            type: "IBAN"
          sepaIndicator: true
          creditor:
            postalAddress:
              unstructured:
                country: "CH"
                addressLines:
                - "Robert Schneider SA"
                - "Rue de la gare 24"
              structured:
                country: "CH"
                streetName: "Rue de la gare"
                townName: "Biel"
                buildingNumber: "24"
                postCode: "2501"
            name: "name"
          remittanceReference:
            reference: "210000000003139471430009017"
            type: "SCOR"
        instructionId: "INSTR-001"
        instructedAmount:
          amount: "10.25"
          currency: "CHF"
        endToEndId: "ENDTOENDID-001"
        isrDetails:
          creditorAccount:
            identification: "762011623852957"
            type: "OTHER"
          creditor:
            postalAddress:
              unstructured:
                country: "CH"
                addressLines:
                - "Robert Schneider SA"
                - "Rue de la gare 24"
              structured:
                country: "CH"
                streetName: "Rue de la gare"
                townName: "Biel"
                buildingNumber: "24"
                postCode: "2501"
            name: "name"
          remittanceReference:
            reference: "210000000003139471430009017"
            type: "ISR"
  paymentInstructionItem:
    type: "object"
    required:
    - "endToEndId"
    - "instructedAmount"
    - "instructionId"
    properties:
      instructionId:
        type: "string"
        example: "INSTR-001"
        maxLength: 35
      endToEndId:
        type: "string"
        example: "ENDTOENDID-001"
        maxLength: 35
      instructedAmount:
        $ref: "#/definitions/paymentCurrencyAmount"
      ibanDetails:
        $ref: "#/definitions/paymentIBANDetail"
      isrDetails:
        $ref: "#/definitions/paymentISRDetail"
      otherDetails:
        $ref: "#/definitions/paymentOtherDetail"
    title: "Payment Instruction Item"
    example:
      otherDetails:
        chargeBearer: "SHAR"
        creditorAgent:
          clearingSystemMemberIdentification:
            code: "CHBCC"
            memberId: "00230"
          bic: "BDEMMXMM"
        creditorAccount:
          identification: "762011623852957"
          type: "OTHER"
        creditor:
          postalAddress:
            unstructured:
              country: "CH"
              addressLines:
              - "Robert Schneider SA"
              - "Rue de la gare 24"
            structured:
              country: "CH"
              streetName: "Rue de la gare"
              townName: "Biel"
              buildingNumber: "24"
              postCode: "2501"
          name: "name"
        remittanceReference:
          reference: "210000000003139471430009017"
          type: "SCOR"
      ibanDetails:
        chargeBearer: "SHAR"
        creditorAgent:
          clearingSystemMemberIdentification:
            code: "CHBCC"
            memberId: "00230"
          bic: "BDEMMXMM"
        remittanceInformation: "Verguetung Juli"
        creditorAccount:
          identification: "CH9300762011623852957"
          type: "IBAN"
        sepaIndicator: true
        creditor:
          postalAddress:
            unstructured:
              country: "CH"
              addressLines:
              - "Robert Schneider SA"
              - "Rue de la gare 24"
            structured:
              country: "CH"
              streetName: "Rue de la gare"
              townName: "Biel"
              buildingNumber: "24"
              postCode: "2501"
          name: "name"
        remittanceReference:
          reference: "210000000003139471430009017"
          type: "SCOR"
      instructionId: "INSTR-001"
      instructedAmount:
        amount: "10.25"
        currency: "CHF"
      endToEndId: "ENDTOENDID-001"
      isrDetails:
        creditorAccount:
          identification: "762011623852957"
          type: "OTHER"
        creditor:
          postalAddress:
            unstructured:
              country: "CH"
              addressLines:
              - "Robert Schneider SA"
              - "Rue de la gare 24"
            structured:
              country: "CH"
              streetName: "Rue de la gare"
              townName: "Biel"
              buildingNumber: "24"
              postCode: "2501"
          name: "name"
        remittanceReference:
          reference: "210000000003139471430009017"
          type: "ISR"
  paymentIBANDetail:
    type: "object"
    required:
    - "creditor"
    - "creditorAccount"
    properties:
      sepaIndicator:
        type: "boolean"
        example: true
      creditorAccount:
        $ref: "#/definitions/paymentIbanAccount"
      creditorAgent:
        $ref: "#/definitions/paymentCreditorAgent"
      creditor:
        $ref: "#/definitions/paymentCreditor"
      chargeBearer:
        $ref: "#/definitions/paymentChargeBearerMethod"
      remittanceReference:
        $ref: "#/definitions/paymentIbanRemittanceReference"
      remittanceInformation:
        type: "string"
        example: "Verguetung Juli"
    title: "Payment IBAN Detail"
    example:
      chargeBearer: "SHAR"
      creditorAgent:
        clearingSystemMemberIdentification:
          code: "CHBCC"
          memberId: "00230"
        bic: "BDEMMXMM"
      remittanceInformation: "Verguetung Juli"
      creditorAccount:
        identification: "CH9300762011623852957"
        type: "IBAN"
      sepaIndicator: true
      creditor:
        postalAddress:
          unstructured:
            country: "CH"
            addressLines:
            - "Robert Schneider SA"
            - "Rue de la gare 24"
          structured:
            country: "CH"
            streetName: "Rue de la gare"
            townName: "Biel"
            buildingNumber: "24"
            postCode: "2501"
        name: "name"
      remittanceReference:
        reference: "210000000003139471430009017"
        type: "SCOR"
...
  paymentBookingInstruction:
    type: "string"
    title: "Payment Booking Instruction"
    example: "BATCHBOOKING_SALA_CND"
    enum:
    - "SINGLEBOOKING_SIA"
    - "SINGLEBOOKING_NOA"
    - "BATCHBOOKING_SALA_NOA"
    - "BATCHBOOKING_SALA_CND"
    - "BATCHBOOKING_NOA"
    - "BATCHBOOKING_CND"
    - "BATCHBOOKING_CWD"
  paymentSubmissionStatus:
    type: "object"
    properties:
      messageId:
        type: "string"
        example: "eb6305c91f7f49deaed016487c27b42d"
      statusCode:
        type: "string"
        example: "PART"
        enum:
        - "ACCP"
        - "RJCT"
        - "PART"
      transactions:
        type: "array"
        items:
          $ref: "#/definitions/paymentInstructionItemStatus"
    title: "Payment Submission Status"
    example:
      messageId: "eb6305c91f7f49deaed016487c27b42d"
      transactions:
      - reasonInformation: "currency USD not allowed for payment type ISR"
        instructionId: "DNCS-20180322-IXN0-TXN0"
        reasonCode: "CURR"
        statusCode: "RJCT"
      - reasonInformation: "currency USD not allowed for payment type ISR"
        instructionId: "DNCS-20180322-IXN0-TXN0"
        reasonCode: "CURR"
        statusCode: "RJCT"
      statusCode: "PART"
  paymentInstructionItemStatus:
    type: "object"
    properties:
      instructionId:
        type: "string"
        example: "DNCS-20180322-IXN0-TXN0"
      statusCode:
        type: "string"
        example: "RJCT"
        enum:
        - "ACCP"
        - "RJCT"
      reasonCode:
        type: "string"
        example: "CURR"
      reasonInformation:
        type: "string"
        example: "currency USD not allowed for payment type ISR"
    title: "Payment Instruction Item Status"
    example:
      reasonInformation: "currency USD not allowed for payment type ISR"
      instructionId: "DNCS-20180322-IXN0-TXN0"
      reasonCode: "CURR"
      statusCode: "RJCT"
  paymentCreditorAgent:
    type: "object"
    properties:
      bic:
        type: "string"
        example: "BDEMMXMM"
      clearingSystemMemberIdentification:
        $ref: "#/definitions/commonClearingSystemMemberIdentification"
    example:
      clearingSystemMemberIdentification:
        code: "CHBCC"
        memberId: "00230"
      bic: "BDEMMXMM"
  paymentIbanAccount:
    type: "object"
    required:
    - "identification"
    - "type"
    properties:
      type:
        type: "string"
        description: "allowed account identification type for the creditor account\
          \ depends on the payment type. The debtor account must always be an IBAN."
        enum:
        - "IBAN"
      identification:
        type: "string"
        example: "CH9300762011623852957"
        maxLength: 34
        pattern: "[A-Z]{2,2}[0-9]{2,2}[a-zA-Z0-9]{1,30}"
    title: "Payment Account"
    example:
      identification: "CH9300762011623852957"
      type: "IBAN"
...
  paymentCurrencyAmount:
    type: "object"
    required:
    - "amount"
    - "currency"
    properties:
      currency:
        type: "string"
        example: "CHF"
      amount:
        type: "string"
        example: "10.25"
        maxLength: 18
        pattern: "[0-9]{1,12}([.][0-9]{1,5})?"
    title: "Payment Currency-Amount"
    example:
      amount: "10.25"
      currency: "CHF"
...
  commonErrorResponse:
    type: "object"
    properties:
      type:
        $ref: "#/definitions/commonErrorType"
      title:
        type: "string"
        example: "This is the general problem description"
      detail:
        type: "string"
        example: "Detailed problem description with respect to the current request,\
          \ e.g., invalid account number format"
      instance:
        type: "string"
        example: "path/to/corresponding/resource"
    title: "Common Error Response"
  commonErrorType:
    type: "string"
    title: "Common Error Type"
    description: "Error Types for commonErrorResponse. See Appendix for detailed information."
    example: "/problems/TECHNICAL_ERROR"
    enum:
    - "/problems/INVALID_PAYLOAD"
    - "/problems/MALFORMED_PAYLOAD"
    - "/problems/INVALID_TOKEN"
    - "/problems/EXPIRED_TOKEN"
    - "/problems/INSUFFICIENT_PRIVILEGES"
    - "/problems/NO_ACCESS_TO_RESOURCE"
    - "/problems/RESOURCE_DOES_NOT_EXIST"
    - "/problems/RESOURCE_NOT_READY"
    - "/problems/RESOURCE_TOO_LARGE"
    - "/problems/WRONG_METHOD"
    - "/problems/OPERATION_NOT_ALLOWED"
    - "/problems/TECHNICAL_ERROR"
...